Fix critical production issues: JWT auth, security, database, and async I/O
This commit addresses 9 critical production-blocking issues identified in the audit:

## User-Requested Features (4 items)
1. **Fix CurrentJob.tsx Progress Display**: Created JobStartedPayload to emit both job and total_frames
2. **Remove Card Gradient Variant**: Removed gradient from Card.tsx and updated usages to glass/solid
3. **Start Paused by Default**: Added auto_start_processing config (defaults false) with UI banner
4. **Remove Old /setup Route**: Deleted Setup.tsx and updated navigation to use /login

## Critical Security Fixes (5 items)
5. **JWT Authentication System**:
   - Added jsonwebtoken dependency and jwt.rs module
   - Implemented token generation/validation (90-day expiration)
   - Created JWT middleware for protected routes
   - Updated client API to send Bearer tokens

6. **Path Traversal Protection**:
   - Created path_security.rs with validation functions
   - Added canonicalization and traversal checks
   - Protected file upload/download endpoints

7. **Job Assignment Race Condition**:
   - Implemented claim_next_job() for atomic get+assign in single transaction
   - Prevents multiple clients from claiming same job

8. **Stale Job Detection Race Condition**:
   - Implemented requeue_stale_jobs() for atomic bulk requeue
   - Updated scheduler to use atomic operation

9. **Transaction Isolation in complete_job()**:
   - Wrapped both encoding_jobs and media_files updates in transaction
   - Ensures atomic completion or rollback

## Critical Bug Fixes
- **Database Schema**: Fixed duration field type from INTEGER to REAL
- **Async I/O Deadlock**: Fixed channel deadlock in frame count extraction by spawning execute task and draining receiver concurrently
- **FFmpeg Output Race**: Changed from tokio::spawn to tokio::join! to start reading stdout/stderr immediately, preventing race with fast processes
- **Incomplete Output Reading**: Switched from line-by-line to read_to_end() to capture all output including chunks without newlines

+++ ffnodes-server/src/media_files/media_file_db.rs
@@ -18,7 +18,7 @@ pub async fn initialize() -> Result<()> {
     size                 INTEGER          DEFAULT NULL,
     scanned_bit_rate     INTEGER NOT NULL,
     bit_rate             INTEGER          DEFAULT NULL,
-    duration             INTEGER NOT NULL,
+    duration             REAL    NOT NULL,
     width                INTEGER NOT NULL,
     height               INTEGER NOT NULL,
     frames               INTEGER NOT NULL,
